requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.28.0
ipython==8.18.1
jedi==0.19.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
MouseInfo==0.1.3
nest-asyncio==1.5.8
openai==1.7.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
PyAutoGUI==0.9.54
pydantic==2.5.3
pydantic_core==2.14.6
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
retry==0.9.2
rpds-py==0.16.2
rubicon-objc==0.4.7
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.13
zipp==3.17.0
